make config integers signed
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2858 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/routers/chaos_router.cpp
@@ -258,7 +258,7 @@ bool ChaosRouter::_InputReady( int input ) const
 
 bool ChaosRouter::_OutputFull( int out ) const
 {
-  return ( _output_frame[out].size( ) >= (unsigned int)_buffer_size );
+  return ( _output_frame[out].size( ) >= (size_t)_buffer_size );
 }
 
 bool ChaosRouter::_OutputAvail( int out ) const
@@ -270,7 +270,7 @@ bool ChaosRouter::_OutputAvail( int out ) const
 
 bool ChaosRouter::_MultiQueueFull( int mq ) const
 {
-  return ( _multi_queue[mq].size( ) >= (unsigned int)_buffer_size );
+  return ( _multi_queue[mq].size( ) >= (size_t)_buffer_size );
 }
 
 int ChaosRouter::_InputForOutput( int output ) const
